WIP
- Moved most of the implementation out of the new example and into an `indexing` submodule of `bevy_ecs`
- Added basic documentation to satisfy linting
