+++ crates/bevy_ecs/src/lib.rs
@@ -10,6 +10,7 @@ pub mod change_detection;
 pub mod component;
 pub mod entity;
 pub mod event;
+pub mod indexing;
 pub mod query;
 #[cfg(feature = "bevy_reflect")]
 pub mod reflect;
